requirements.txt
beautifulsoup4
firebase-admin
get-gecko-driver
google-api-python-client
gTTS
objdict_bf
odfpy
openai
PyPDF2
python-dotenv
python-docx
requests
seleniumbase
spotipy
streamlit
streamlit-input-box
streamlit-mic-recorder
streamlit-TTS
streamlit-pdf-reader
streamlit-modal-input
tiktoken
matplotlib
numpy
pandas
graphviz
altair
plotly
bokeh
pydeck
scipy
sympy
scikit-learn
vega-datasets
